fix(OMN-12700): satisfy infra validation invariants
