+++ src/omnibase_infra/nodes/node_llm_inference_effect/models/model_llm_call_completed_event.py
@@ -1,15 +1,13 @@
 # SPDX-FileCopyrightText: 2025 OmniNode.ai Inc.
 # SPDX-License-Identifier: MIT
-"""LLM call completion events emitted by the inference effect node."""
+"""Full LLM call completion event emitted by the inference effect node."""
 
 from __future__ import annotations
 
 from pydantic import BaseModel, ConfigDict, Field
 
 from omnibase_core.types import JsonType
-
-TOPIC_LLM_CALL_COMPLETED = "onex.evt.omniintelligence.llm-call-completed.v1"
-TOPIC_LLM_CALL_COMPLETED_INFRA = "onex.evt.omnibase-infra.llm-call-completed.v1"
+from omnibase_infra.topics import SUFFIX_INTELLIGENCE_LLM_CALL_COMPLETED
 
 
 class ModelLlmCallCompletedEvent(BaseModel):
@@ -17,9 +15,10 @@ class ModelLlmCallCompletedEvent(BaseModel):
 
     model_config = ConfigDict(frozen=True, extra="forbid", from_attributes=True)
 
-    topic: str = TOPIC_LLM_CALL_COMPLETED
+    topic: str = SUFFIX_INTELLIGENCE_LLM_CALL_COMPLETED
     schema_version: str = "1.0"
-    model_id: str
+    # ONEX_EXCLUDE: pattern_validator - model_id is an LLM model name (e.g. "gemini-2.5-pro"), not a UUID entity reference
+    model_id: str = Field(min_length=1)
     prompt_tokens: int = 0
     completion_tokens: int = 0
     total_tokens: int = 0
@@ -40,29 +39,4 @@ class ModelLlmCallCompletedEvent(BaseModel):
     compute_usage_source: str | None = None
 
 
-class ModelLlmCallCompletedInfraEvent(BaseModel):
-    """Infra-local LLM completion event for infra projections."""
-
-    model_config = ConfigDict(frozen=True, extra="forbid", from_attributes=True)
-
-    topic: str = TOPIC_LLM_CALL_COMPLETED_INFRA
-    model_id: str
-    endpoint_url: str
-    prompt_tokens: int = 0
-    completion_tokens: int = 0
-    total_tokens: int = 0
-    latency_ms: float
-    success: bool = True
-    timestamp: str
-    gpu_seconds: float | None = None
-    gpu_type: str | None = None
-    gpu_count: int | None = None
-    compute_usage_source: str | None = None
-
-
-__all__: list[str] = [
-    "ModelLlmCallCompletedEvent",
-    "ModelLlmCallCompletedInfraEvent",
-    "TOPIC_LLM_CALL_COMPLETED",
-    "TOPIC_LLM_CALL_COMPLETED_INFRA",
-]
+__all__: list[str] = ["ModelLlmCallCompletedEvent"]
