+++ src/omnibase_infra/validation/validation_exemptions.yaml
@@ -91,6 +91,32 @@ skip_directories:
 # These handle method count, parameter count, naming, and style violations
 pattern_exemptions:
   # ==========================================================================
+  # LLM Metrics Event Model Exemptions (OMN-12700)
+  # ==========================================================================
+  # These event payloads preserve the existing llm-call-completed wire contract.
+  # Here model_id is an LLM model name such as "gemini-2.5-pro", not an ONEX
+  # entity identifier. Cost/projection consumers already read this field by key.
+  - file_pattern: 'model_llm_call_completed_event\.py'
+    violation_pattern: "Field 'model_id' should use UUID type instead of str"
+    reason: >
+      The omniintelligence llm-call-completed metrics event uses model_id for the
+      provider/model name. Changing this field to UUID would break the existing
+      metrics payload contract and downstream cost aggregation consumers.
+
+    documentation:
+      - OMN-12700 PR description
+    ticket: OMN-12700
+  - file_pattern: 'model_llm_call_completed_infra_event\.py'
+    violation_pattern: "Field 'model_id' should use UUID type instead of str"
+    reason: >
+      The infra-local llm-call-completed event mirrors the metrics model_id as an
+      LLM provider/model name. It is not an ONEX entity identifier and must stay a
+      string for compatibility with existing infra projections.
+
+    documentation:
+      - OMN-12700 PR description
+    ticket: OMN-12700
+  # ==========================================================================
   # ContractLinter Exemptions (OMN-517)
   # ==========================================================================
   # Contract linter with cohesive validation methods, each validating a
